+++ HISTORY.rst
@@ -3,6 +3,16 @@
 History
 -------
 
+0.5.2 (2016-02-??)
+~~~~~~~~~~~~~~~~~~
+
+* Capture separate metrics counting phase 1 snps (varscan) and phase 2 snps (consensus). Previously, the
+  metrics only included phase 1 snps.  This changes the contents of both the ``metrics`` and ``metrics.tsv``
+  files. The metrics file now contains a new tag ``phase1Snps``.  The old tag ``snps`` now correctly counts 
+  the total number of snps. The metrics.tsv file now has separate column headers for phase 1 snps and 
+  phase 2 snps.  Any code that parses those files may need modifications to work properly with v0.5.2.
+
+
 0.5.1 (2016-02-19)
 ~~~~~~~~~~~~~~~~~~